@@ -1,10 +1,11 @@
 module github.com/hknutzen/Netspoc-Approve/go
 
-go 1.22.1
+go 1.23.1
 
 require (
-	github.com/google/go-cmp v0.5.9
-	github.com/hknutzen/testtxt v0.0.0-20240305140207-e6d67e884e6e
+	github.com/creack/pty v1.1.23
+	github.com/google/go-cmp v0.6.0
+	github.com/hknutzen/testtxt v0.0.0-20240408182449-0168fe18ebfb
 	github.com/pkg/diff v0.0.0-20210226163009-20ebb0f2a09e
 	github.com/spf13/pflag v1.0.5
 	github.com/tailscale/goexpect v0.0.0-20210902213824-6e8c725cea41
@@ -14,7 +15,6 @@ require (
 )
 
 require (
-	github.com/creack/pty v1.1.23 // indirect
 	github.com/google/goterm v0.0.0-20200907032337-555d40f16ae2 // indirect
 	golang.org/x/crypto v0.17.0 // indirect
 	golang.org/x/sys v0.15.0 // indirect
