@@ -1,27 +1,28 @@
 package device
 
 import (
-	"crypto/tls"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/fs"
-	"net"
-	"net/http"
-	"net/url"
 	"os"
 	"path"
-	"strings"
 	"syscall"
-	"time"
 
-	"github.com/spf13/pflag"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/asa"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/codefiles"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/deviceconf"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/errlog"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/ios"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/linux"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/nsx"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/panos"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/program"
 )
 
 type RealDevice interface {
-	LoadDevice(fname string, c *Config, l1, l2 *os.File) (DeviceConfig, error)
-	ParseConfig(data []byte, fName string) (DeviceConfig, error)
-	GetChanges(c1, c2 DeviceConfig) error
+	LoadDevice(fname string, c *program.Config, l1, l2 *os.File) (
+		deviceconf.Config, error)
+	ParseConfig(data []byte, fName string) (deviceconf.Config, error)
+	GetChanges(c1, c2 deviceconf.Config) error
 	GetErrUnmanaged() []error
 	ApplyCommands(*os.File) error
 	HasChanges() bool
@@ -29,74 +30,78 @@ type RealDevice interface {
 	CloseConnection()
 }
 
-type DeviceConfig interface {
-	MergeSpoc(DeviceConfig) DeviceConfig
+func getRealDevice(fname string) RealDevice {
+	info, _ := codefiles.LoadInfoFile(fname)
+	switch info.Model {
+	case "ASA":
+		return asa.Setup()
+	case "IOS":
+		return ios.Setup()
+	case "Linux":
+		return &linux.State{}
+	case "NSX":
+		return &nsx.State{}
+	case "PAN-OS":
+		return &panos.State{}
+	default:
+		errlog.Abort("Unexpected model %q in file %s.info\n",
+			info.Model, fname)
+	}
+	return nil
 }
 
 type state struct {
 	RealDevice
-	config   *Config
+	config   *program.Config
 	logFname string
 }
 
-var quiet bool
-
-func Main(dev RealDevice, fs *pflag.FlagSet) int {
-	s := &state{RealDevice: dev}
-	getString := func(fs *pflag.FlagSet, name string) string {
-		val, _ := fs.GetString(name)
-		return val
-	}
-	quiet, _ = fs.GetBool("quiet")
-	return handleBailout(func() int {
-		var err error
-		args := fs.Args()
-		switch len(args) {
-		case 2:
-			q := fs.Changed("quiet")
-			n := fs.NFlag()
-			if q && n > 1 || !q && n > 0 {
-				fs.Usage()
-				return 1
-			}
-			s.setStderrLog("")
-			err = s.compareFiles(args[0], args[1])
-		case 1:
-			fname := args[0]
-			s.config, err = LoadConfig()
-			if err != nil {
-				break
-			}
-			s.config.User = getString(fs, "user")
-			s.setLogDir(getString(fs, "logdir"), fname)
-			s.setStderrLog(getString(fs, "LOGFILE"))
-			s.setLock(fname)
-			if v, _ := fs.GetBool("compare"); v {
-				err = s.compare(fname)
-			} else {
-				err = s.approve(fname)
-			}
-			s.CloseConnection()
+func ApproveOrCompare(
+	isCompare bool,
+	fname string,
+	cfg *program.Config,
+	logDir string,
+	logFile string,
+	quiet bool,
+) int {
+	return errlog.HandleAbort(func() int {
+		errlog.Quiet = quiet
+		errlog.SetStderrLog(logFile)
+		s := &state{RealDevice: getRealDevice(fname)}
+		s.config = cfg
+		if logDir != "" {
+			s.setLogDir(logDir, fname)
 		}
+		var err error
+		if isCompare {
+			err = s.compare(fname)
+		} else {
+			err = s.approve(fname)
+		}
+		s.CloseConnection()
 		if err != nil {
-			Abort("%v", err)
+			errlog.Abort("%v", err)
 		}
 		return 0
 	})
 }
 
-func (s *state) compareFiles(fname1, fname2 string) error {
-	conf1, err := s.loadSpoc(fname1)
-	if err != nil {
-		return err
-	}
-	err = s.getCompare(conf1, fname2)
-	if err != nil {
-		return err
-	}
-	s.showCompareInfo()
-	fmt.Print(s.ShowChanges())
-	return nil
+func CompareFiles(fname1, fname2 string, quiet bool) int {
+	return errlog.HandleAbort(func() int {
+		errlog.Quiet = quiet
+		errlog.SetStderrLog("")
+		s := &state{RealDevice: getRealDevice(fname2)}
+		conf1, err := s.loadSpoc(fname1)
+		if err != nil {
+			errlog.Abort("%v", err)
+		}
+		if err := s.getCompare(conf1, fname2); err != nil {
+			errlog.Abort("%v", err)
+		}
+		s.showCompareInfo()
+		fmt.Print(s.ShowChanges())
+		return 0
+	})
 }
 
 func (s *state) compare(fname string) error {
@@ -105,7 +110,7 @@ func (s *state) compare(fname string) error {
 		return err
 	}
 	for _, w := range s.GetErrUnmanaged() {
-		Warning("%v", w)
+		errlog.Warning("%v", w)
 	}
 	s.showCompareInfo()
 	if s.logFname != "" && s.HasChanges() {
@@ -138,7 +143,7 @@ func (s *state) compareDevice(fname string) error {
 	return s.getCompare(conf1, fname)
 }
 
-func (s *state) loadDevice(fname string) (DeviceConfig, error) {
+func (s *state) loadDevice(fname string) (deviceconf.Config, error) {
 	logConfig, err := s.getLogFH(".config")
 	if err != nil {
 		return nil, err
@@ -159,7 +164,7 @@ func (s *state) applyCommands() error {
 	}
 	defer closeLogFH(logFH)
 	if !s.HasChanges() {
-		DoLog(logFH, "No changes applied")
+		errlog.DoLog(logFH, "No changes applied")
 		return nil
 	}
 	return s.ApplyCommands(logFH)
@@ -167,13 +172,13 @@ func (s *state) applyCommands() error {
 
 func (s *state) showCompareInfo() {
 	if !s.HasChanges() {
-		Info("comp: device unchanged")
+		errlog.Info("comp: device unchanged")
 	} else {
-		Info("comp: *** device changed ***")
+		errlog.Info("comp: *** device changed ***")
 	}
 }
 
-func (s *state) getCompare(c1 DeviceConfig, fname string) error {
+func (s *state) getCompare(c1 deviceconf.Config, fname string) error {
 	c2, err := s.loadSpoc(fname)
 	if err != nil {
 		return err
@@ -181,8 +186,8 @@ func (s *state) getCompare(c1 DeviceConfig, fname string) error {
 	return s.GetChanges(c1, c2)
 }
 
-func (s *state) loadSpoc(v4Path string) (DeviceConfig, error) {
-	v6Path := getIPv6Fname(v4Path)
+func (s *state) loadSpoc(v4Path string) (deviceconf.Config, error) {
+	v6Path := codefiles.GetIPv6Fname(v4Path)
 	conf4, err := s.loadSpocFile(v4Path)
 	if err != nil {
 		return nil, err
@@ -195,7 +200,7 @@ func (s *state) loadSpoc(v4Path string) (DeviceConfig, error) {
 	return s.addRaw(conf, v4Path)
 }
 
-func (s *state) addRaw(conf DeviceConfig, v4Path string) (DeviceConfig, error) {
+func (s *state) addRaw(conf deviceconf.Config, v4Path string) (deviceconf.Config, error) {
 	rawPath := v4Path + ".raw"
 	raw, err := s.loadSpocFile(rawPath)
 	if err != nil {
@@ -205,7 +210,7 @@ func (s *state) addRaw(conf DeviceConfig, v4Path string) (DeviceConfig, error) {
 	return conf, nil
 }
 
-func (s *state) loadSpocFile(fname string) (DeviceConfig, error) {
+func (s *state) loadSpocFile(fname string) (deviceconf.Config, error) {
 	data, err := os.ReadFile(fname)
 	if err != nil && !errors.Is(err, os.ErrNotExist) {
 		return nil, fmt.Errorf("Can't %v", err)
@@ -218,97 +223,18 @@ func (s *state) loadSpocFile(fname string) (DeviceConfig, error) {
 	return c, nil
 }
 
-func getIPv6Fname(p string) string {
-	dir := path.Dir(p)
-	base := path.Base(p)
-	return dir + "/ipv6/" + base
-}
-
 // Set lock for exclusive approval.
-// Store file handle in global var, so it isn't closed immediately.
-// File is closed automatically after program exit.
-var lockFH *os.File
-
-func (s *state) setLock(fname string) {
-	lockFile := path.Join(s.config.lockfileDir, path.Base(fname))
-	_, statErr := os.Stat(lockFile)
-	fh, err := os.Create(lockFile)
+func SetLock(fname, dir string) (*os.File, error) {
+	lockFile := path.Join(dir, path.Base(fname))
+	fh, err := os.OpenFile(lockFile, os.O_CREATE|os.O_RDONLY, 0666)
 	if err != nil {
-		Abort("Can't %v", err)
-	}
-	// Make newly created lock file writable for other users.
-	if statErr != nil && errors.Is(statErr, fs.ErrNotExist) {
-		os.Chmod(lockFile, 0666)
+		return nil, err
 	}
 	err = syscall.Flock(int(fh.Fd()), syscall.LOCK_EX|syscall.LOCK_NB)
 	if err != nil {
-		Abort("Approve in progress for %s", fname)
+		err = fmt.Errorf("Approve in progress for %s", fname)
 	}
-	lockFH = fh
-}
-
-type codeInfo struct {
-	GeneratedBy             string   `json:"generated_by"`
-	Model                   string   `json:"model"`
-	IPList                  []string `json:"ip_list,omitempty"`
-	NameList                []string `json:"name_list,omitempty"`
-	PolicyDistributionPoint string   `json:"policy_distribution_point,omitempty"`
-}
-
-func getHostnameIPList(path string) ([]string, []string, error) {
-	info, checked := LoadInfoFile(path)
-	nameList := info.NameList
-	ipList := info.IPList
-	if len(nameList) == 0 {
-		return nil, nil, fmt.Errorf("Missing device name in %v", checked)
-	}
-	if len(ipList) == 0 {
-		return nil, nil, fmt.Errorf("Missing IP address in %v", checked)
-	}
-	if len(nameList) != len(ipList) {
-		return nil, nil, fmt.Errorf(
-			"Number of device names and IP addresses don't match in %v", checked)
-	}
-	return nameList, ipList, nil
-}
-
-func GetIPPDP(fName string) (string, string, error) {
-	info, checked := LoadInfoFile(fName)
-	ipList := info.IPList
-	if len(ipList) == 0 {
-		return "", "", fmt.Errorf("Missing IP address in %v", checked)
-	}
-	return ipList[0], info.PolicyDistributionPoint, nil
-}
-
-func GetHostname(fName string) string {
-	return path.Base(fName)
-}
-
-func LoadInfoFile(path string) (*codeInfo, []string) {
-	path6 := getIPv6Fname(path)
-	info := &codeInfo{}
-	var checked []string
-	for _, file := range []string{path, path6} {
-		file += ".info"
-		fd, err := os.Open(file)
-		if err != nil {
-			if errors.Is(err, os.ErrNotExist) {
-				continue
-			}
-			panic(err)
-		}
-		checked = append(checked, file)
-		defer fd.Close()
-		if err := json.NewDecoder(fd).Decode(&info); err != nil {
-			panic(err)
-		}
-		// Must also read IPv6 file if v4 file has no IP.
-		if len(info.IPList) > 0 {
-			break
-		}
-	}
-	return info, checked
+	return fh, err
 }
 
 func (s *state) setLogDir(logDir, file string) {
@@ -322,124 +248,12 @@ func (s *state) getLogFH(ext string) (*os.File, error) {
 		return nil, nil
 	}
 	fname := s.logFname + ext
-	moveLogFile(fname)
-	return createWithPath(fname)
-}
-
-// Rename existing logfile.
-func moveLogFile(fname string) {
-	if _, err := os.Stat(fname); err == nil {
-		os.Rename(fname, fmt.Sprintf("%s.%d", fname, time.Now().Unix()))
-	}
+	errlog.MoveLogFile(fname)
+	return errlog.CreateWithPath(fname)
 }
 
 func closeLogFH(fh *os.File) {
 	if fh != nil {
 		fh.Close()
 	}
-}
-
-func DoLog(fh *os.File, s string) {
-	if fh != nil {
-		if strings.HasPrefix(s, "http") || strings.HasPrefix(s, "action=") {
-			s, _ = url.QueryUnescape(s)
-		}
-		fmt.Fprintln(fh, s)
-	}
-}
-
-var stderrLog *os.File
-
-func (s *state) setStderrLog(fname string) {
-	stderrLog = os.Stderr
-	if fname != "" {
-		moveLogFile(fname)
-		fh, err := createWithPath(fname)
-		if err != nil {
-			Abort("Can't %v", err)
-		}
-		stderrLog = fh
-	}
-}
-
-func createWithPath(fname string) (*os.File, error) {
-	dir := path.Dir(fname)
-	err := os.MkdirAll(dir, 0755)
-	if err != nil {
-		return nil, err
-	}
-	return os.Create(fname)
-}
-
-func GetHTTPClient(cfg *Config) *http.Client {
-	return &http.Client{
-		Timeout: time.Duration(cfg.Timeout) * time.Second,
-		Transport: &http.Transport{
-			Dial: (&net.Dialer{
-				// Set low connection timeout, so we can better switch to
-				// backup device.
-				Timeout: time.Duration(cfg.LoginTimeout) * time.Second,
-			}).Dial,
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		},
-	}
-}
-
-func TryReachableHTTPLogin(
-	fname string,
-	cfg *Config,
-	login func(name, ip, user, pass string) error,
-) error {
-
-	nameList, ipList, err := getHostnameIPList(fname)
-	if err != nil {
-		return err
-	}
-	for i, name := range nameList {
-		ip := ipList[i]
-		user, pass := cfg.GetUserPass(name)
-		if err := login(name, ip, user, pass); err != nil {
-			Warning("%v", err)
-			continue
-		}
-		return nil
-	}
-	return fmt.Errorf(
-		"Devices unreachable: %s", strings.Join(nameList, ", "))
-}
-
-func Info(format string, args ...interface{}) {
-	if !quiet {
-		fmt.Fprintf(stderrLog, format+"\n", args...)
-	}
-}
-
-func Warning(format string, args ...interface{}) {
-	printWithMarker("WARNING>>> ", format, args...)
-}
-
-type bailout struct{}
-
-func handleBailout(f func() int) (exitCode int) {
-	defer func() {
-		if e := recover(); e != nil {
-			if _, ok := e.(bailout); !ok {
-				panic(e) // Resume same panic if it's not a bailout.
-			}
-			exitCode = 1
-		}
-	}()
-	return f()
-}
-
-func Abort(format string, args ...interface{}) {
-	printWithMarker("ERROR>>> ", format, args...)
-	panic(bailout{})
-}
-
-func printWithMarker(m string, format string, args ...interface{}) {
-	out := fmt.Sprintf(format, args...)
-	out = strings.TrimSuffix(out, "\n")
-	out = strings.ReplaceAll(out, "\n", "\n"+m)
-	fmt.Fprintln(stderrLog, m+out)
 }
