@@ -2,7 +2,6 @@ package ios
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"os"
 	"regexp"
@@ -10,14 +9,15 @@ import (
 	"time"
 
 	"github.com/hknutzen/Netspoc-Approve/go/pkg/cisco"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/codefiles"
 	"github.com/hknutzen/Netspoc-Approve/go/pkg/console"
-	"github.com/hknutzen/Netspoc-Approve/go/pkg/device"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/deviceconf"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/errlog"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/program"
 )
 
 type State struct {
 	cisco.State
-	conn         *console.Conn
-	errUnmanaged []error
 	reloadActive bool
 }
 
@@ -29,135 +29,82 @@ func Setup() *State {
 }
 
 func (s *State) ParseConfig(data []byte, fName string) (
-	device.DeviceConfig, error) {
+	deviceconf.Config, error) {
 
 	return s.State.ParseConfig(removeBanner(data), fName)
 }
 
 func (s *State) LoadDevice(
-	spocFile string, cfg *device.Config, logLogin, logConfig *os.File) (
-	device.DeviceConfig, error) {
+	spocFile string, cfg *program.Config, logLogin, logConfig *os.File) (
+	deviceconf.Config, error) {
 
-	user, pass := cfg.GetUserPass(device.GetHostname(spocFile))
-	var err error
-	s.conn, err = console.GetSSHConn(spocFile, user, cfg, logLogin)
+	user, pass, err := cfg.GetUserPass(codefiles.GetHostname(spocFile))
 	if err != nil {
 		return nil, err
 	}
-	hostName := device.GetHostname(spocFile)
-	s.loginEnable(pass, cfg)
+	s.Conn, err = console.GetSSHConn(spocFile, user, cfg, logLogin)
+	if err != nil {
+		return nil, err
+	}
+	hostName := codefiles.GetHostname(spocFile)
+	s.LoginEnable(pass, cfg)
 	s.setTerminal()
 	s.logVersion()
 	s.checkDeviceName(hostName)
 
-	s.conn.SetLogFH(logConfig)
-	device.Info("Requesting device config")
-	out := s.conn.GetCmdOutput("sh run")
-	device.Info("Got device config")
+	s.Conn.SetLogFH(logConfig)
+	errlog.Info("Requesting device config")
+	out := s.Conn.GetCmdOutput("sh run")
+	errlog.Info("Got device config")
 	config, err := s.ParseConfig([]byte(out), "<device>")
-	device.Info("Parsed device config")
+	errlog.Info("Parsed device config")
 	if err != nil {
 		err = fmt.Errorf("While reading device: %v", err)
 	}
 	return config, err
 }
 
-func (s *State) loginEnable(pass string, cfg *device.Config) {
-	var bannerLines string
-	conn := s.conn
-	out := conn.ShortWait(`(?i)password:|\(yes/no.*\)\?`)
-	if strings.HasSuffix(out, "?") {
-		out = conn.IssueCmd("yes", `(?i)password:`)
-	}
-	bannerLines += out
-	// Look for prompt. Ignore prompt lines with whitespace or multiple
-	// hash that may occur in lines of banner.
-	waitPrompt := func(enter, suffix string) bool {
-		stdPrompt := `\r\n\r?[^#> ]+[>#] ?$`
-		out = conn.IssueCmd(enter, `(?i)password:|`+stdPrompt)
-		bannerLines += out
-		out = strings.TrimSuffix(out, " ")
-		return strings.HasSuffix(out, suffix)
-	}
-	if waitPrompt(pass, ">") {
-		// Enter enable mode.
-		if !waitPrompt("enable", "#") {
-			// Enable password required.
-			// Use login password as enable password.
-			if !waitPrompt(pass, "#") {
-				device.Abort("Authentication for enable mode failed")
-			}
-		}
-	} else if !strings.HasSuffix(out, "#") {
-		device.Abort("Authentication failed")
-	}
-
-	// Force new prompt by issuing empty command.
-	// Use this prompt because of performance impact of standard prompt.
-	out = conn.IssueCmd("", `#[ ]?`)
-	i := strings.LastIndex(out, "\n")
-	// Current prompt: "\n\rHOSTNAME# "
-	p := out[i:]
-	// Prompt may vary before terminating "# "
-	i = strings.LastIndex(p, "#")
-	rx := regexp.MustCompile(
-		regexp.QuoteMeta(p[:i]) + `\S*` + regexp.QuoteMeta(p[i:]))
-	conn.SetStdPrompt(rx)
-	s.checkBanner(bannerLines, cfg)
-}
-
 func (s *State) setTerminal() {
-	s.conn.SendCmd("term len 0")
-	s.conn.SendCmd("term width 512")
+	s.Conn.SendCmd("term len 0")
+	s.Conn.SendCmd("term width 512")
 }
 
 func (s *State) logVersion() {
-	s.conn.GetCmdOutput("sh ver")
+	s.Conn.GetCmdOutput("sh ver")
 }
 
 func (s *State) checkDeviceName(name string) {
 	// Force new prompt by issuing empty command.
 	// Output is: \r\n\s*NAME#\s?
-	out := strings.TrimSpace(s.conn.IssueCmd("", `#[ ]?`))
+	out := strings.TrimSpace(s.Conn.IssueCmd("", `#[ ]?`))
 	out = strings.TrimSuffix(out, "#")
 	if name != out {
-		device.Abort("Wrong device name: %q, expected: %q", out, name)
+		errlog.Abort("Wrong device name: %q, expected: %q", out, name)
 	}
-}
-
-func (s *State) checkBanner(lines string, cfg *device.Config) {
-	if rx := cfg.CheckBanner; rx != nil && rx.FindStringIndex(lines) == nil {
-		s.errUnmanaged =
-			[]error{errors.New("Missing banner at NetSPoC managed device")}
-	}
-}
-
-func (s *State) GetErrUnmanaged() []error {
-	return s.errUnmanaged
 }
 
 func (s *State) prepareDevice() {
-	s.conn.SendCmd("configure terminal")
+	s.Conn.SendCmd("configure terminal")
 	// Don't slow down the system by logging to console.
-	s.conn.SendCmd("no logging console")
+	s.Conn.SendCmd("no logging console")
 	// Enable logging synchronous to get a fresh prompt after
 	// a reload banner is shown.
-	s.conn.SendCmd("line vty 0 15")
-	s.conn.SendCmd("logging synchronous level all")
+	s.Conn.SendCmd("line vty 0 15")
+	s.Conn.SendCmd("logging synchronous level all")
 	// Needed for default route to work as expected.
-	s.conn.SendCmd("ip subnet-zero")
-	s.conn.SendCmd("ip classless")
-	s.conn.SendCmd("end")
+	s.Conn.SendCmd("ip subnet-zero")
+	s.Conn.SendCmd("ip classless")
+	s.Conn.SendCmd("end")
 }
 
 func (s *State) ApplyCommands(logFh *os.File) error {
-	s.conn.SetLogFH(logFh)
+	s.Conn.SetLogFH(logFh)
 	s.prepareDevice()
 	func() {
 		s.scheduleReload()
 		defer s.cancelReload()
-		s.conn.SendCmd("configure terminal")
-		defer s.conn.SendCmd("end")
+		s.Conn.SendCmd("configure terminal")
+		defer s.Conn.SendCmd("end")
 		for _, chg := range s.Changes {
 			s.cmd(chg)
 		}
@@ -186,9 +133,9 @@ func (s *State) ApplyCommands(logFh *os.File) error {
 func (s *State) writeMem() {
 	retries := 2
 	for {
-		out := s.conn.IssueCmd("write memory", `#[ ]?|\[confirm\]`)
+		out := s.Conn.IssueCmd("write memory", `#[ ]?|\[confirm\]`)
 		if strings.Contains(out, "Overwrite the previous NVRAM configuration") {
-			out = s.conn.GetCmdOutput("")
+			out = s.Conn.GetCmdOutput("")
 		}
 		if strings.Contains(out, "[OK]") {
 			return
@@ -199,9 +146,9 @@ func (s *State) writeMem() {
 				time.Sleep(3 * time.Second)
 				continue
 			}
-			device.Abort("write mem: startup-config open failed - giving up")
+			errlog.Abort("write mem: startup-config open failed - giving up")
 		}
-		device.Abort("write mem: unexpected result: %s", out)
+		errlog.Abort("write mem: unexpected result: %s", out)
 	}
 }
 
@@ -209,15 +156,15 @@ func (s *State) writeMem() {
 // No output expected from commands.
 func (s *State) cmd(cmd string) {
 	c1, c2, _ := strings.Cut(cmd, "\n")
-	s.conn.Send(cmd)
+	s.Conn.Send(cmd)
 	needReload := false
 	check := func(ci string) {
-		out := s.conn.GetOutput()
+		out := s.Conn.GetOutput()
 		out, needReload = s.stripReloadBanner(out)
-		out = s.conn.StripEcho(ci, out)
+		out = s.Conn.StripEcho(ci, out)
 		if out != "" {
 			if !isValidOutput(ci, out) {
-				device.Abort("Got unexpected output from '%s':\n%s", ci, out)
+				errlog.Abort("Got unexpected output from '%s':\n%s", ci, out)
 			}
 		}
 	}
@@ -239,7 +186,7 @@ func isValidOutput(cmd, out string) bool {
 			continue
 		}
 		if strings.HasPrefix(line, "WARNING:") {
-			device.Warning("Got unexpected output from '%s':\n%s", cmd, line)
+			errlog.Warning("Got unexpected output from '%s':\n%s", cmd, line)
 			continue
 		}
 		return false
@@ -248,8 +195,8 @@ func isValidOutput(cmd, out string) bool {
 }
 
 func (s *State) CloseConnection() {
-	if c := s.conn; c != nil {
-		s.conn.Close()
+	if c := s.Conn; c != nil {
+		s.Conn.Close()
 	}
 }
 
@@ -268,26 +215,26 @@ func (s *State) sendReloadCmd(withDo bool) {
 	if withDo {
 		cmd = "do " + cmd
 	}
-	out := s.conn.IssueCmd(cmd, `\[yes\/no\]:\ |\[confirm\]`)
+	out := s.Conn.IssueCmd(cmd, `\[yes\/no\]:\ |\[confirm\]`)
 	// System configuration has been modified. Save? [yes/no]:
 	if strings.Contains(out, "[yes/no]") {
 		// Leave our changes unsaved, to be sure that a reload
 		// gets last good configuration.
-		s.conn.IssueCmd("n", `\[confirm\]`)
+		s.Conn.IssueCmd("n", `\[confirm\]`)
 	}
 	// Confirm the reload with empty command, wait for the standard prompt.
 	s.reloadActive = true
-	s.conn.SendCmd("")
+	s.Conn.SendCmd("")
 }
 
 func (s *State) cancelReload() {
 	// Don't wait for standard prompt, but for banner message, which is
 	// sent asynchronously.
-	s.conn.IssueCmd("reload cancel", `--- SHUTDOWN ABORTED ---`)
+	s.Conn.IssueCmd("reload cancel", `--- SHUTDOWN ABORTED ---`)
 	// Because of 'logging synchronous' we are sure to get another prompt.
-	s.conn.ShortWait(`[#] ?$`)
+	s.Conn.WaitShort(`[#] ?$`)
 	// Synchronize expect buffers with empty command.
-	s.conn.SendCmd("")
+	s.Conn.SendCmd("")
 	s.reloadActive = false
 }
 
@@ -322,15 +269,15 @@ func (s *State) stripReloadBanner(out string) (string, bool) {
 				// Because of 'logging synchronous' we are sure to get another prompt
 				// if the banner is the only output before current prompt.
 				// Read next prompt.
-				device.Info("Found banner before output, expecting another prompt")
-				out = s.conn.ShortWait(`[#] ?$`)
-				out = s.conn.StripStdPrompt(out)
+				errlog.Info("Found banner before output, expecting another prompt")
+				out = s.Conn.WaitShort(`[#] ?$`)
+				out = s.Conn.StripStdPrompt(out)
 			} else if prefix != "" && strings.TrimSpace(postfix) == "" {
 				// Try to read another prompt if banner is shown directly
 				// behind current output.
-				device.Info("Found banner after output, checking another prompt")
-				if s.conn.TryPrompt() {
-					device.Info("- Found prompt")
+				errlog.Info("Found banner after output, checking another prompt")
+				if s.Conn.TryPrompt() {
+					errlog.Info("- Found prompt")
 				}
 			}
 			matched, _ := regexp.MatchString(`SHUTDOWN in 0?0:01:00`, msg)
