@@ -3,29 +3,30 @@ package linux
 import (
 	"slices"
 
-	"github.com/hknutzen/Netspoc-Approve/go/pkg/device"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/deviceconf"
+	"github.com/hknutzen/Netspoc-Approve/go/pkg/errlog"
 )
 
-func (a *config) MergeSpoc(d device.DeviceConfig) device.DeviceConfig {
+func (a *config) MergeSpoc(d deviceconf.Config) deviceconf.Config {
 	b := d.(*config)
 	a.routes = append(a.routes, b.routes...)
 	for tName, bChains := range b.iptables {
 		aChains := a.iptables[tName]
 		if aChains == nil {
-			device.Info("Adding all chains of table %q", tName)
+			errlog.Info("Adding all chains of table %q", tName)
 			a.iptables[tName] = bChains
 			continue
 		}
 		for cName, bChain := range bChains {
 			aChain := aChains[cName]
 			if aChain == nil {
-				device.Info("Adding chain %q of table %q", cName, tName)
+				errlog.Info("Adding chain %q of table %q", cName, tName)
 				aChains[cName] = bChain
 				continue
 			}
 			switch aChain.policy {
 			case "-", "":
-				device.Abort("Must not redefine chain %q of table %q from rawdata",
+				errlog.Abort("Must not redefine chain %q of table %q from rawdata",
 					cName, tName)
 			}
 			for _, ru := range bChain.rules {
